todo comments added and WIP action data
the big question is how to do resolution of actions:
if the view is to pass data on the action chosen to the manager, how do we handle tying animations to resolutions of the action's events, like element swapping or health-change numbers?

Since actions are likely just going to be IEnumerators, I'm thinking we'll just have the function reference itself passed into the Manager to be run as a Coroutine, calling View and Model changes as it goes. We'll see how I feel tomorrow, though.